Removes protocols
