--- FakeNFT/Services/AlertFactory/AlertFactoryProtocol.swift
@@ -1,5 +1,0 @@
-import UIKit
-
-protocol AlertFactoryProtocol {
-    func makeSortingAlert() -> UIAlertController
-}
