+++ FakeNFT/Services/ServiceFactory/ServiceFactory.swift
@@ -1,5 +1,9 @@
 import Foundation
 
+protocol ServiceFactoryProtocol {
+    func makeAlertFactory() -> AlertFactoryProtocol
+}
+
 final class ServiceFactory: ServiceFactoryProtocol {
     func makeAlertFactory() -> AlertFactoryProtocol {
         AlertFactory()
