--- FakeNFT/Services/ServiceFactory/ServiceFactoryProtocol.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-protocol ServiceFactoryProtocol {
-    func makeAlertFactory() -> AlertFactoryProtocol
-}
